#20 feat(OctaveTestActivity, OctaveTestSingingActivity): separate octaveTest and octaveTestSinging Activity
